Collect aliases in Strive.Aliases
