+++ library/Strive/Aliases.hs
@@ -1,7 +1,24 @@
 -- | Aliases for common types.
-module Strive.Aliases
-  ( Result
-  ) where
+module Strive.Aliases where
 
--- | An action's result.
+import Data.Time.Clock (UTCTime)
+
+type ActivityId = Integer
+type ApplicationId = Integer
+type ApplicationSecret = String
+type AthleteId = Integer
+type AuthorizationCode = String
+type ClubId = Integer
+type EffortId = Integer
+type ElapsedTime = Integer
+type Extension = String
+type GearId = String
+type Latitude = Double
+type Longitude = Double
+type Name = String
+type RedirectUri = String
 type Result a = IO (Either String a)
+type SegmentId = Integer
+type StartTime = UTCTime
+type StreamId = Integer
+type UploadId = Integer
